Finished separating routes and got the endpoints working. Added isAuthenticated function to restricted endpoints. It's not really global middleware at this point though.
