+++ data/dbConfig.js
@@ -3,3 +3,5 @@ const knex = require('knex');
 const knexConfig = require('../knexfile.js');
 
 module.exports = knex(knexConfig.development);
+
+
