+++ index.js
@@ -1,27 +1,14 @@
 const express = require('express');
-const session = require('express-session');
 const cors = require('cors');
-const bcrypt = require('bcryptjs');
 
-const db = require('./data/dbConfig.js');
 
-const authRoutes = require('./auth_helpers/authRoutes.js');
+// const db = require('./data/dbConfig.js');
+
+const authRoutes = require('./authHelpers/authRoutes.js');
 
 const server = express();
 
-const sessionConfig = {
-    secret: 'nobody-tosses.a%dwarf.!',
-    name: 'user-session',
-    httpOnly: true,
-    resave: false,
-    saveUninitialized: false,
-    cookie: {
-      secure: false,
-      maxAge: 1000 * 60 * 1
-    },
-};
 
-server.use(session(sessionConfig));
 server.use(express.json());
 server.use(cors());
